backend/requirements.txt
# PropManage Backend — Python Dependencies
# Core
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
pydantic[email]==2.6.3
httpx==0.27.0
aiofiles==23.2.1

# Supabase
supabase==2.4.0

# AI / ML
google-generativeai==0.4.0
anthropic==0.18.0
openai==1.12.0
tiktoken==0.6.0

# Scheduling
apscheduler==3.10.4

# Notifications
twilio==9.0.0
sendgrid==6.11.0

# PDF Processing
PyMuPDF==1.23.22

# Utils
python-dotenv==1.0.1
python-dateutil==2.9.0
